backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.82.0
replicate==0.15.4
Pillow==10.0.1
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
reportlab==4.0.4
ezdxf==1.0.3
cairosvg==2.7.1
numpy==1.24.3
opencv-python-headless==4.8.0.76
opencv-contrib-python-headless==4.8.0.76 